requirements.txt
Django==3.1.7
gunicorn==20.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
colorlog==5.0.1
djangorestframework-simplejwt==4.7.2